[창희] 총 강좌 리스트 CRUD 및 CSS 수정 / 삭제 시 FK로 잡혀있어 SQL Exception이 뜰 경우 방지해 이전에 FK로 잡혀있는 페이지 확인 후 alert 경고문 띄움

+++ LMS/src/main/java/com/gd/LMS/lecture/controller/TotalLectureController.java
@@ -1,15 +1,16 @@
 package com.gd.LMS.lecture.controller;
 
+import java.sql.SQLException;
 import java.util.List;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.gd.LMS.commons.TeamColor;
@@ -24,32 +25,60 @@ import lombok.extern.slf4j.Slf4j;
 public class TotalLectureController {
 	@Autowired TotalLectureService totalLectureService;
 	
-	// 직원 전체과목 보기
+	// 전체과목 보기 (직원기능)
 	@GetMapping("/employee/totalLectureList")
-	public String totalLectureList(PagingVo vo, Model model,
+	public String totalLectureList(PagingVo pagingVo, Model model, Map<String, Object> map,
 			@RequestParam(value = "currentPage", defaultValue = "1") int currentPage,
 			@RequestParam(value = "rowPerPage", defaultValue = "10") int rowPerPage,
-			@RequestParam(value = "keyword", defaultValue = "") String keyword) {
-		int totalLectureCount = totalLectureService.getTotalLectureCount(keyword);
-		vo = new PagingVo(currentPage, totalLectureCount, rowPerPage, keyword, null);
-
+			@RequestParam(value = "keyword", defaultValue = "") String keyword,
+			@RequestParam(value = "searchType", defaultValue = "") String searchType) {
+		// 파라미터 설정
+		map.put("keyword", keyword);
+		map.put("searchType", searchType);
+		log.debug(TeamColor.LCH + "map > " + map);
+		
+		// 전체 강의 개수 및 페이징 변수 vo에 추가 후 디버깅
+		int totalLectureCount = totalLectureService.getTotalLectureCount(map);
 		log.debug(TeamColor.LCH + "count > " + totalLectureCount);
-		log.debug(TeamColor.LCH + "vo > " + vo);
-
-		List<Map<String, Object>> totalLectureList = totalLectureService.getTotalLectureList(vo);
-
-		model.addAttribute("paging", vo);
+		
+		// 이전 페이지 시작 글 번호와 현재 변경되는 페이지의 시작 글번호에 대한 일치 시키는거 많은 변경이 필요하므로 그냥 1로 처리함
+		pagingVo = new PagingVo(currentPage, totalLectureCount, rowPerPage, keyword, searchType);
+		if (pagingVo.getBeginRow() >= totalLectureCount) {
+			pagingVo = new PagingVo(1, totalLectureCount, rowPerPage, keyword, searchType);
+		}
+		log.debug(TeamColor.KJS + "PagingVo : " + pagingVo);
+		
+		// 전체 부서이름 리스트에 담은 후 디버깅
+		List<Map<String, Object>> list = totalLectureService.getTotalDepartmentName();
+		log.debug(TeamColor.LCH + "list > " + list);
+		
+		// 리스트에 강의목록 담기
+		List<Map<String, Object>> totalLectureList = totalLectureService.getTotalLectureList(pagingVo);
+		
+		// 모델에 담아서 페이지로 리턴
+		model.addAttribute("department", list);
+		model.addAttribute("paging", pagingVo);
 		model.addAttribute("totalList", totalLectureList);
 
 		return "lecture/totalLecture/totalLectureList";
 	}
 	
 	@GetMapping("/employee/totalLectureOne")
-	public String totalLectureOne(Model model, @RequestParam (value = "lectureCode") int lectureCode) {
+	public String totalLectureOne(Model model
+			, @RequestParam (value = "errorMsg", defaultValue = "") String errorMsg
+			, @RequestParam (value = "lectureCode") int lectureCode) {
 		log.debug(TeamColor.LCH + "--- totalLectureOne Controller GetMapping ---");
 		
 		Map<String, Object> map = totalLectureService.getTotalLectureOne(lectureCode);
 		log.debug(TeamColor.LCH + "totalLectureOne > " + map);
+		
+		// error 처리
+		if(!errorMsg.equals("")) {
+			log.debug(TeamColor.LCH + "errorMsg > " + errorMsg);
+			model.addAttribute("errorMsg", errorMsg);
+		} else {
+			model.asMap().clear();
+		}
 		
 		model.addAttribute("map", map);
 		
@@ -65,7 +94,7 @@ public class TotalLectureController {
 		
 		model.addAttribute("department", list);
 		
-		return "lecture/totalLecture/addTotalLectureOne";
+		return "lecture/totalLecture/addTotalLecture";
 	}
 	
 	@PostMapping("/employee/addTotalLecture")
@@ -78,7 +107,7 @@ public class TotalLectureController {
 		log.debug(TeamColor.LCH + "result > " + result);
 		
 		redirectAttributes.addAttribute("lectureCode", result);
-		return "redirect:/totalLectureOne";
+		return "redirect:/employee/totalLectureOne";
 	}
 	
 	@GetMapping("/employee/modifyTotalLecture")
@@ -107,19 +136,30 @@ public class TotalLectureController {
 		log.debug(TeamColor.LCH + "result > " + result);
 		
 		redirectAttributes.addAttribute("lectureCode", totalLecture.getLectureCode());
-		return "redirect:/totalLectureOne";
+		return "redirect:/employee/totalLectureOne";
 	}
 	
 	@GetMapping("/employee/removeTotalLecture")
-	public String removeTotalLecture(@RequestParam (value = "lectureCode") int lectureCode) {
+	public String removeTotalLecture(Model model, RedirectAttributes redirectAttributes
+			, @RequestParam (value = "lectureCode") int lectureCode) {
 		log.debug(TeamColor.LCH + "--- removeTotalLecture Controller GetMapping ---");
 		
 		log.debug(TeamColor.LCH + "lectureCode > " + lectureCode);
 		
+		// FK로 잡힌 테이블이 있는지 먼저 확인해야함
+		List<String> checkFk = totalLectureService.getCheckFkOpenedLecture(lectureCode);
+		log.debug(TeamColor.LCH + "checkFk 0? > " + checkFk.size());
+		if(checkFk.size() != 0) {
+			log.debug(TeamColor.LCH + "삭제실패 FK 존재");
+			redirectAttributes.addAttribute("lectureCode", lectureCode);
+			redirectAttributes.addAttribute("errorMsg", "진행중인 강의가 있어 삭제가 불가능합니다.");
+			return "redirect:/employee/totalLectureOne";
+		}
+		
 		int result = totalLectureService.removeTotalLectureOne(lectureCode);
 		log.debug(TeamColor.LCH + "result > " + result);
 		
-		return "redirect:/totalLectureList";
+		return "redirect:/employee/totalLectureList";
 	}
 	
 }
